requirements.txt
transformers
sentence-transformers
faiss-cpu
pandas
requests
torch
accelerate
python-dotenv
